requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-mysql==3.8.1
django-object-actions==2.0.0
djangorestframework==3.12.2
glad==0.1.33
gprof2dot==2019.11.30
idna==2.10
mysqlclient==2.0.1
oauthlib==3.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
python3-openid==3.2.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.25.11
uWSGI==2.0.19.1
Pillow==8.1.0